refactoring: add pipelines/kernel folder (#2875)
