+++ cpp/open3d/t/pipelines/kernel/TransformationConverter.cpp
@@ -24,16 +24,17 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/t/pipelines/TransformationConverter.h"
+#include "open3d/t/pipelines/kernel/TransformationConverter.h"
 
 #include <cmath>
 
 #include "open3d/core/Tensor.h"
-#include "open3d/t/pipelines/TransformationConverterImpl.h"
+#include "open3d/t/pipelines/kernel/TransformationConverterImpl.h"
 
 namespace open3d {
 namespace t {
 namespace pipelines {
+namespace kernel {
 
 core::Tensor RtToTransformation(const core::Tensor &R, const core::Tensor &t) {
     core::Dtype dtype = core::Dtype::Float32;
@@ -92,6 +93,7 @@ core::Tensor PoseToTransformation(const core::Tensor &pose) {
     return transformation;
 }
 
+}  // namespace kernel
 }  // namespace pipelines
 }  // namespace t
 }  // namespace open3d
