+++ cpp/open3d/t/pipelines/kernel/TransformationConverter.cu
@@ -27,11 +27,12 @@
 #include <cuda.h>
 #include <cuda_runtime.h>
 
-#include "open3d/t/pipelines/TransformationConverterImpl.h"
+#include "open3d/t/pipelines/kernel/TransformationConverterImpl.h"
 
 namespace open3d {
 namespace t {
 namespace pipelines {
+namespace kernel {
 
 __global__ void PoseToTransformationKernel(float *transformation_ptr,
                                            const float *X_ptr) {
@@ -42,6 +43,7 @@ void PoseToTransformationCUDA(float *transformation_ptr, const float *X_ptr) {
     PoseToTransformationKernel<<<1, 1>>>(transformation_ptr, X_ptr);
 }
 
+}  // namespace kernel
 }  // namespace pipelines
 }  // namespace t
 }  // namespace open3d
